+++ .azure/app-cloud-e2e.yml
@@ -31,7 +31,7 @@ pr:
       - ".azure/app-cloud-e2e.yml"
       - "src/lightning_app/**"
       - "requirements/app/**"
-      - "tests/tests_examples_app/**"
+      - "tests/integrations_app_examples/**"
       - "setup.py"
     exclude:
       - "requirements/*/docs.txt"
@@ -98,6 +98,17 @@ jobs:
     # values: https://docs.microsoft.com/en-us/azure/devops/pipelines/process/phases?view=azure-devops&tabs=yaml#workspace
     workspace:
       clean: all
+    variables:
+      HEADLESS: '1'
+      PACKAGE_LIGHTNING: '1'
+      CLOUD: '1'
+      VIDEO_LOCATION: $(video_artifact_dir)
+      PR_NUMBER: $(local_id)
+      TEST_APP_NAME: $(name)
+      TEST_APP_FOLDER: $(dir)
+      HAR_LOCATION: './artifacts/hars'
+      SLOW_MO: '50'
+      LIGHTNING_DEBUG: '1'
     steps:
 
     - script: echo '##vso[task.setvariable variable=local_id]$(System.PullRequest.PullRequestNumber)'
@@ -137,7 +148,7 @@ jobs:
     # - bash: |
     #     rm -rf examples/app_template_jupyterlab || true
     #     git clone https://github.com/Lightning-AI/LAI-lightning-template-jupyterlab-App examples/app_template_jupyterlab
-    #     cp examples/app_template_jupyterlab/tests/test_template_jupyterlab.py tests/tests_examples_app/test_template_jupyterlab.py
+    #     cp examples/app_template_jupyterlab/tests/test_template_jupyterlab.py tests/integrations_app_examples/test_template_jupyterlab.py
     #   condition: eq(variables['name'], 'template_jupyterlab')
     #   displayName: 'Clone Template Jupyter Lab Repo'
 
@@ -165,26 +176,17 @@ jobs:
     - bash: |
         mkdir -p ${VIDEO_LOCATION}
         ls -l examples/${TEST_APP_NAME}
-        ls -l tests/tests_examples_app/public
-        python -m pytest tests/tests_examples_app/${TEST_APP_FOLDER}/test_${TEST_APP_NAME}.py::test_${TEST_APP_NAME}_example_cloud \
+        ls -l tests/integrations_app_examples/public
+        python -m pytest ${TEST_FILE}::test_${TEST_APP_NAME}_example_cloud \
           --timeout=540 --capture=no -v --color=yes
       env:
-        HEADLESS: '1'
-        PACKAGE_LIGHTNING: '1'
-        CLOUD: '1'
-        VIDEO_LOCATION: $(video_artifact_dir)
-        PR_NUMBER: $(local_id)
-        TEST_APP_NAME: $(name)
-        TEST_APP_FOLDER: $(dir)
-        HAR_LOCATION: './artifacts/hars'
-        SLOW_MO: '50'
+        TEST_FILE: tests/integrations_app_examples/$(TEST_APP_FOLDER)/test_$(TEST_APP_NAME).py
         #LAI_USER: $(LAI_USER)  # for STAGING
         #LAI_PASS: $(LAI_PASS)  # for STAGING
         LIGHTNING_USER_ID: $(LIGHTNING_USER_ID_PROD)
         LIGHTNING_API_KEY: $(LIGHTNING_API_KEY_PROD)
         LIGHTNING_USERNAME: $(LIGHTNING_USERNAME)
         LIGHTNING_CLOUD_URL: $(LIGHTNING_CLOUD_URL_PROD)
-        LIGHTNING_DEBUG: '1'
       displayName: 'Run the tests'
 
     - task: PublishPipelineArtifact@1
@@ -204,11 +206,5 @@ jobs:
         LIGHTNING_API_KEY: $(LIGHTNING_API_KEY_PROD)
         LIGHTNING_USERNAME: $(LIGHTNING_USERNAME)
         LIGHTNING_CLOUD_URL: $(LIGHTNING_CLOUD_URL_PROD)
-        PR_NUMBER: $(local_id)
-        TEST_APP_NAME: $(name)
-        TEST_APP_FOLDER: $(dir)
-        # GRID_USER_ID: $(LIGHTNING_USER_ID)  # TODO: clarify the meaning
-        # GRID_USER_KEY: $(LIGHTNING_API_KEY)  # TODO: clarify the meaning
-        # GRID_URL: $(LIGHTNING_CLOUD_URL)
-        # _GRID_USERNAME: $(LIGHTNING_USERNAME)
+      timeoutInMinutes: "3"
       displayName: 'Clean Previous Apps'
