+++ .github/checkgroup.yml
@@ -247,7 +247,7 @@ subprojects:
       - ".actions/**"
       - ".github/workflows/ci-examples-app.yml"
       - "src/lightning_app/**"
-      - "tests/tests_examples_app/**"
+      - "tests/integrations_app_examples/**"
       - "examples/app_*/**"
       - "requirements/app/**"
       - "setup.py"
@@ -271,7 +271,7 @@ subprojects:
       - ".azure/app-cloud-e2e.yml"
       - "src/lightning_app/**"
       - "requirements/app/**"
-      - "tests/tests_examples_app/**"
+      - "tests/integrations_app_examples/**"
       - "setup.py"
       - "!requirements/*/docs.txt"
       - "!*.md"
