+++ .github/labeler.yml
@@ -5,7 +5,7 @@ app:
   all: ['!src/pytorch_lightning/**/*', '!tests/tests_pytorch/**/*', '!docs/source-pytorch/**/*', '!requirements/pytorch/*']
 - any: ['tests/tests_app/**/*']
   all: ['!src/pytorch_lightning/**/*', '!tests/tests_pytorch/**/*', '!docs/source-pytorch/**/*', '!requirements/pytorch/*']
-- any: ['tests/tests_examples_app/**/*']
+- any: ['tests/integrations_app_examples/**/*']
   all: ['!src/pytorch_lightning/**/*', '!tests/tests_pytorch/**/*', '!docs/source-pytorch/**/*', '!requirements/pytorch/*']
 - any: ['examples/app_*/**/*']
   all: ['!src/pytorch_lightning/**/*', '!tests/tests_pytorch/**/*', '!docs/source-pytorch/**/*', '!requirements/pytorch/*']
@@ -16,12 +16,12 @@ app:
 
 pl:
 - any: ['src/pytorch_lightning/**/*']
-  all: ['!src/lightning/**/*', '!src/lightning_app/**/*', '!tests/tests_app/**/*', '!tests/tests_examples_app/**/*', '!docs/source-app/**/*', '!requirements/app/*']
+  all: ['!src/lightning/**/*', '!src/lightning_app/**/*', '!tests/tests_app/**/*', '!tests/integrations_app_examples/**/*', '!docs/source-app/**/*', '!requirements/app/*']
 - any: ['examples/*pl_*/**/*']
-  all: ['!src/lightning/**/*', '!src/lightning_app/**/*', '!tests/tests_app/**/*', '!tests/tests_examples_app/**/*', '!docs/source-app/**/*', '!requirements/app/*']
+  all: ['!src/lightning/**/*', '!src/lightning_app/**/*', '!tests/tests_app/**/*', '!tests/integrations_app_examples/**/*', '!docs/source-app/**/*', '!requirements/app/*']
 - any: ['tests/tests_pytorch/**/*']
-  all: ['!src/lightning/**/*', '!src/lightning_app/**/*', '!tests/tests_app/**/*', '!tests/tests_examples_app/**/*', '!docs/source-app/**/*', '!requirements/app/*']
+  all: ['!src/lightning/**/*', '!src/lightning_app/**/*', '!tests/tests_app/**/*', '!tests/integrations_app_examples/**/*', '!docs/source-app/**/*', '!requirements/app/*']
 - any: ['docs/source-pytorch/**/*']
-  all: ['!src/lightning/**/*', '!src/lightning_app/**/*', '!tests/tests_app/**/*', '!tests/tests_examples_app/**/*', '!docs/source-app/**/*', '!requirements/app/*']
+  all: ['!src/lightning/**/*', '!src/lightning_app/**/*', '!tests/tests_app/**/*', '!tests/integrations_app_examples/**/*', '!docs/source-app/**/*', '!requirements/app/*']
 - any: ['requirements/pytorch/*']
-  all: ['!src/lightning/**/*', '!src/lightning_app/**/*', '!tests/tests_app/**/*', '!tests/tests_examples_app/**/*', '!docs/source-app/**/*', '!requirements/app/*']
+  all: ['!src/lightning/**/*', '!src/lightning_app/**/*', '!tests/tests_app/**/*', '!tests/integrations_app_examples/**/*', '!docs/source-app/**/*', '!requirements/app/*']
