+++ tests/integrations_app_examples/apps/collect_failures/__init__.py
