+++ tests/integrations_app_examples/apps/collect_failures/app.py
