+++ tests/integrations_app_examples/apps/collect_failures/requirements.txt
