+++ tests/integrations_app_examples/apps/core_features_app/__init__.py
