+++ tests/integrations_app_examples/apps/core_features_app/app.py
@@ -8,7 +8,7 @@ class EnvVarTestApp(LightningFlow):
         super().__init__()
 
     def run(self):
-        # these env vars are set here: tests/tests_examples_app/test_core_features_app.py:15
+        # these env vars are set here: tests/integrations_app_examples/test_core_features_app.py:15
         assert os.getenv("FOO", "") == "bar"
         assert os.getenv("BLA", "") == "bloz"
         self._exit()
