+++ tests/integrations_app_examples/apps/custom_work_dependencies/app.py
