+++ tests/integrations_app_examples/local/test_custom_work_dependencies.py
@@ -2,7 +2,7 @@ import os
 from time import sleep
 
 import pytest
-from tests_examples_app.local import _PATH_APPS
+from integrations_app_examples.local import _PATH_APPS
 
 from lightning_app.testing.testing import run_app_in_cloud
