+++ tests/integrations_app_examples/public/test_argparse.py
@@ -1,7 +1,7 @@
 import os
 import sys
 
-from tests_examples_app.public import _PATH_EXAMPLES
+from integrations_app_examples.public import _PATH_EXAMPLES
 
 from lightning_app.testing.testing import application_testing
 from lightning_app.utilities.load_app import _patch_sys_argv
