+++ tests/integrations_app_examples/public/test_template_streamlit_ui.py
@@ -2,7 +2,7 @@ import os
 from time import sleep
 
 import pytest
-from tests_examples_app.public import _PATH_EXAMPLES
+from integrations_app_examples.public import _PATH_EXAMPLES
 
 from lightning_app.testing.testing import run_app_in_cloud, wait_for
