+++ tests/integrations_app_examples/public/test_v0_app.py
@@ -5,7 +5,7 @@ from unittest import mock
 from unittest.mock import MagicMock
 
 import pytest
-from tests_examples_app.public import _PATH_EXAMPLES
+from integrations_app_examples.public import _PATH_EXAMPLES
 
 from lightning_app import LightningApp
 from lightning_app.runners import CloudRuntime
